pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentboot"
version = "0.0.1"
description = "AI agent for bare-metal hardware bootstrapping and OS installation"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.10"
authors = [
    { name = "Francisco Angulo de Lafuente", email = "agnuxo1@gmail.com" },
]
keywords = ["ai-agent", "llm", "bare-metal", "bootstrapping", "hardware", "sysadmin"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: System Administrators",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Installation/Setup",
    "Topic :: System :: Systems Administration",
]
dependencies = [
    "llama-cpp-python>=0.3.20",
    "huggingface-hub>=0.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[project.scripts]
agentboot = "agentboot.cli:main"

[project.urls]
Homepage = "https://github.com/Agnuxo1/AgentBoot"
Repository = "https://github.com/Agnuxo1/AgentBoot"
Issues = "https://github.com/Agnuxo1/AgentBoot/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
